Split component template and component class tests into different files
