+++ tests/test_component.py
@@ -1,0 +1,37 @@
+import json
+import os
+
+from pathlib import Path as P
+
+from git import Repo
+
+from commodore.component import Component, component_dir
+from commodore.inventory import Inventory
+
+
+def setup_directory(tmp_path: P):
+    os.chdir(tmp_path)
+
+    inv = Inventory(work_dir=tmp_path)
+    inv.ensure_dirs()
+
+    jsonnetfile = tmp_path / "jsonnetfile.json"
+    with open(jsonnetfile, "w") as jf:
+        json.dump({"version": 1, "dependencies": [], "legacyImports": True}, jf)
+
+
+def _init_repo(tmp_path: P, cn: str, url: str):
+    setup_directory(tmp_path)
+    cr = Repo.init(component_dir(cn))
+    cr.create_remote("origin", url)
+
+
+def test_init_existing_component(tmp_path: P):
+    cn = "test-component"
+    orig_url = "git@github.com:projectsyn/commodore.git"
+    _init_repo(tmp_path, cn, orig_url)
+
+    c = Component(cn)
+
+    for url in c.repo.remote().urls:
+        assert url == orig_url
